Create simplecalculator.py
Simple calculator which will perform the simplest mathematical operations with numbers sequentially, accepting operands (numbers) and an operator from the user.

Conditions for this task

The application works with integers and real numbers.
The application can perform the following mathematical operations:
ADD (+)
SUBTRACTION(-)
MULTIPLICATION (*)
DIVISION (/)
The program accepts one operand or one statement in one request-response cycle.
The program performs all operations in the order of arrival - one by one.
The program displays the result of calculations when it receives the symbol = from the user.
The application ends after it outputs the result of the calculation.
The user alternately enters numbers and operators.
If the user enters the operator twice in a row, they receive an error message and can re-enter.
If the user enters a number twice in a row, they receive an error message and can enter again.
The application correctly processes the situation of incorrect input (exception).